Added breakpoints and hamburger
Added breakpoint for mobileonly to disable nav-links and display the hamburger. Downloaded and imported Tasty css hamburgers by jonsuh and imported the button to the header. Now need to link up to javascript to add class: 'is-active' for it to start the animation. Project: Added hover animation to increase opacity of image.